feat: implement ActiveReservationPage with reservation details, services, and products management
